Update of the script, now with Shape-File-Ability!
Update of the script, now with Shape-File-Ability!

Conversion of the training an validation data in tif files is not necessary anymore!!
You can directly link your training and validation classes as shape files.
Furthermore, the classification result will be masked. Thus a black border has a value of 0 in the classification.